requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
confluent-kafka==2.5.3
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
celery
redis
elastic-transport==8.15.0
elasticsearch==8.15.1
email_validator==2.2.0
exceptiongroup==1.2.2
facenet-pytorch==2.6.0
fastapi==0.112.0
fastapi-cli==0.0.5
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
gremlinpython==3.7.2
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.7
importlib_resources==6.4.4
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
lz4==4.3.3
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.0.5
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pillow==10.2.0
protobuf==4.25.5
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-louvain==0.16
python-multipart==0.0.10
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
scipy==1.14.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.37.2
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.1
opt_einsum==3.4.0
langchain
openai
psycopg2-binary
python-dotenv
python-jose
langchain_openai
passlib
bcrypt